Move HttpError to separate dir
