+++ src/scripts/utils/errors/HttpError.js
